requirements.txt
flask
pika
pytest
pytest-sugar
pytest-progress
pytest-instafail
pytest-testrail
docker
requests
mysql-connector-python
influxdb_client